[alerts] refactor alert types to be versionable
