+++ nexus/alert-types/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "nexus-alert-types"
+version = "0.1.0"
+edition.workspace = true
+license = "MPL-2.0"
+
+[lints]
+workspace = true
+
+[features]
+test-utils = ["dep:expectorate", "dep:camino"]
+
+[build-dependencies]
+omicron-rpaths.workspace = true
+
+[dependencies]
+camino = { workspace = true, optional = true }
+expectorate = { workspace = true, optional = true }
+schemars.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+serde_with.workspace = true
+strum.workspace = true
+
+[dev-dependencies]
+expectorate.workspace = true
+camino.workspace = true
