+++ nexus/types/Cargo.toml
@@ -61,6 +61,7 @@ trust-quorum-types.workspace = true
 tufaceous-artifact.workspace = true
 
 newtype-uuid.workspace = true
+nexus-alert-types.workspace = true
 nexus-types-versions.workspace = true
 oxide-update-engine.workspace = true
 oxide-update-engine-types.workspace = true
