+++ nexus/types/versions/Cargo.toml
@@ -18,6 +18,7 @@ http.workspace = true
 lldp_protocol.workspace = true
 mg-admin-client.workspace = true
 mg-api-types.workspace = true
+nexus-alert-types.workspace = true
 omicron-common.workspace = true
 omicron-passwords.workspace = true
 omicron-uuid-kinds.workspace = true
@@ -40,3 +41,6 @@ tough.workspace = true
 tufaceous-artifact.workspace = true
 url.workspace = true
 uuid.workspace = true
+
+[dev-dependencies]
+nexus-alert-types = { workspace = true, features = ["test-utils"] }
