+++ nexus/types/versions/src/alert_list/mod.rs
@@ -1,0 +1,10 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+//! Version `ALERT_LSIT` of the Nexus external API.
+//!
+//! Adds an `alert_list` endpoint, and brings the existing alert types into the
+//! versioned API.
+
+pub mod alert;
