+++ nexus/types/versions/src/latest.rs
@@ -56,6 +56,18 @@ pub mod alert {
     pub use crate::v2025_11_20_00::alert::WebhookSecretCreate;
     pub use crate::v2025_11_20_00::alert::WebhookSecretSelector;
     pub use crate::v2025_11_20_00::alert::WebhookSecrets;
+    pub use crate::v2026_08_11_00::alert::Alert;
+
+    // Alert class submodules
+    pub mod power_shelf {
+        pub use crate::v2026_08_11_00::alert::power_shelf::PowerShelf;
+        pub use crate::v2026_08_11_00::alert::power_shelf::Psu;
+        pub use crate::v2026_08_11_00::alert::power_shelf::PsuIdentity;
+        pub use crate::v2026_08_11_00::alert::power_shelf::PsuInsertedV0;
+        pub use crate::v2026_08_11_00::alert::power_shelf::PsuInsertedVersions;
+        pub use crate::v2026_08_11_00::alert::power_shelf::PsuRemovedV0;
+        pub use crate::v2026_08_11_00::alert::power_shelf::PsuRemovedVersions;
+    }
 }
 
 pub mod audit {
