+++ nexus/types/versions/src/lib.rs
@@ -97,3 +97,5 @@ pub mod v2026_06_08_00;
 pub mod v2026_06_10_00;
 #[path = "add_system_ip_pool_apis/mod.rs"]
 pub mod v2026_06_11_00;
+#[path = "alert_list/mod.rs"]
+pub mod v2026_08_11_00;
